changes are done for size file
