+++ src/stories/size.js
@@ -1,0 +1,34 @@
+const size = {
+  h1: {
+    fontSize: '3rem',
+    lineHeight: '3.6rem', 
+    fontWeight: '800',
+  },
+  h2: {
+    fontSize: '2.5rem',
+    lineHeight: '3rem', 
+    fontWeight: '700',
+  },
+  h3: {
+    fontSize: '1.8rem',
+    lineHeight: '2.16rem', 
+    fontWeight: '600',
+  },
+  h4: {
+    fontSize: '1.4rem',
+    lineHeight: '1.68rem', 
+    fontWeight: '500',
+  },
+  h5: {
+    fontSize: '1.2rem',
+    lineHeight: '1.44rem', 
+    fontWeight: '500',
+  },
+  h6: {
+    fontSize: '1rem',
+    lineHeight: '1.2rem', 
+    fontWeight: '400',
+  },
+};
+
+export default size;
